Fix doc push error
Trying to push a JSON document with identical contents to `test/fixtures/doc.json` was failing, due to a "missing_id" error.  This is because it is looking for the `_id` property on the `doc` object - however, `doc` is actually a wrapper object with a property called `doc` that contains the actual document.  The fix uses the nested document to check for the presence of _id.