feat: diff user docs
user docs have special fields, `derived_key`, `iterations`, `password_scheme`
and `salt` which are calculated server side based on the `password` property.
This change ignores those fields for comparison as well as the password field as
long as the password has not been changed.
This prevents couchdb-push to update user docs even if they have not changed.
The resulting doc will have a different `salt` and `derived_key` and existing
sessions will be invalid.

Closes #25